requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
DateTime==5.1
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.4
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
termcolor==2.2.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
zope.interface==6.0
